batch size resolve2

+++ training/unsup/simclr_CSI.py
@@ -70,7 +70,7 @@ def train(P, epoch, model, criterion, optimizer, scheduler, loader, train_exposu
 
         # Creating shift labels to match combined batch size
         shift_labels = torch.cat([torch.ones(batch_size, device=device), torch.zeros(batch_size, device=device)])
-        shift_labels = shift_labels.repeat(2)
+        shift_labels = shift_labels.repeat(2).long()
 
         images_pair = torch.cat([images1, images2], dim=0)  # 8B
         images_pair = simclr_aug(images_pair)  # transform
